refactor: split host internal routes
